% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversions.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to Fahrenheit conversion}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Kelvin}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Kelvin temperatures to degrees Fahrenheit
}
\examples{
temp1 <- K_to_F(50);
temp2 <- K_to_F( c(50, 63, 23) );
}
